Filter only the top 500 data points
This changes index.{html,haml} to use the full data.csv. This is
untenable because of the size, so we just filter the first 500 largest
values.

It's still a bit janky and hard to make heads or tails of the data, but
it's at least a pretty good starting point!